voice-agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
strands-agents[openai]
strands-agents-tools
pyyaml
websockets==13.1
python-multipart==0.0.12
